Entities paging (#3475)
* Check index manager is running

* Handle case no default entity ID is specified

* #3471 Use magma paging when navigating in value sets

+++ opal-search/src/main/java/org/obiba/opal/search/service/OpalSearchService.java
@@ -266,6 +266,7 @@ public class OpalSearchService implements Service {
 
   @Subscribe
   public void onVariablesUpdated(VariablesUpdatedEvent event) {
+    if (!isRunning()) return;
     // to ensure variable search is correct
     getVariablesIndexManager().getIndex(event.getValueTable()).delete();
     // synchronize variable index
